+++ shuup/themes/classic_gray/build_resources.js
@@ -1,0 +1,22 @@
+/**
+ * This file is part of Shuup.
+ *
+ * Copyright (c) 2012-2018, Shuup Inc. All rights reserved.
+ *
+ * This source code is licensed under the OSL-3.0 license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+const { getParcelBuildCommand, runBuildCommands } = require("shuup-static-build-tools");
+
+runBuildCommands([
+    getParcelBuildCommand({
+        cacheDir: "classic_gray",
+        outputDir: "static/shuup/classic_gray/pink",
+        entryFile: "static_src/pink/style.js"
+    }),
+    getParcelBuildCommand({
+        cacheDir: "classic_gray",
+        outputDir: "static/shuup/classic_gray/blue",
+        entryFile: "static_src/blue/style.js"
+    })
+]);